requirements.txt
Flask>=3.1,<4
mysql-connector-python>=9,<10
python-dotenv>=1.0,<2
gunicorn>=21,<24
waitress>=3,<4
